functional protoype with youtube as a module
